backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
httpx>=0.24.0
psycopg2-binary>=2.9.5
asyncpg>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
